Refactor logic of disabled states, and centralize quality checks.

+++ app/src/main/java/org/wikipedia/dataclient/mwapi/MwQueryPage.java
@@ -39,7 +39,7 @@ public class MwQueryPage extends BaseModel {
     @Nullable private String convertedFrom;
     @Nullable private String convertedTo;
     @Nullable private Map<String, String> varianttitles;
-    @SerializedName("pageviews") @Nullable private Map<String, Integer> pageViewsMap;
+    @SerializedName("pageviews") @Nullable private Map<String, Long> pageViewsMap;
 
     @NonNull public String title() {
         return StringUtils.defaultString(title);
@@ -139,7 +139,7 @@ public class MwQueryPage extends BaseModel {
         return varianttitles != null ? StringUtils.defaultIfEmpty(varianttitles.get(langCode), title()) : title();
     }
 
-    @NonNull public Map<String, Integer> getPageViewsMap() {
+    @NonNull public Map<String, Long> getPageViewsMap() {
         return pageViewsMap != null ? pageViewsMap : Collections.emptyMap();
     }
